refactor: extract normalizePhone to tools/jobs/phone.ts
Importing send-weekly.ts in tests ran its top-level main().catch() block, crashing with missing GMAIL_USER secret. Moved normalizePhone into its own module so tests don't drag in the CLI entry point. send-weekly re-exports for backwards compat.
